+++ src/main/java/com/PickOne/domain/user/service/UserService.java
@@ -18,6 +18,12 @@ public class UserService {
                 .orElseThrow(() -> new IllegalArgumentException("해당 사용자를 찾을 수 없습니다. ID=" + id));
     }
 
+    @Transactional(readOnly = true)
+    public User findByEmail(String email) {
+        return userRepository.findByEmail(email)
+                .orElseThrow(() -> new IllegalArgumentException("해당 이메일을 가진 사용자를 찾을 수 없습니다. email=" + email));
+    }
+
     @Transactional
     public User updateUser(Long id, User updateData) {
         User user = findById(id);
