+++ src/main/java/com/PickOne/global/verification/dto/VerificationResponse.java
@@ -1,4 +1,4 @@
-package com.PickOne.global.auth.dto;
+package com.PickOne.global.verification.dto;
 
 public record VerificationResponse(
         boolean success,
