+++ src/main/java/com/PickOne/global/verification/event/EmailVerifiedEvent.java
@@ -1,4 +1,4 @@
-package com.PickOne.global.auth.event;
+package com.PickOne.global.verification.event;
 
 import lombok.Getter;
 import org.springframework.context.ApplicationEvent;
